feat(client): wire refund button and info into OrderDetailsModal.
