+++ client/src/components/pages/Store/Orders/OrderDetailsModal/index.js
@@ -1,0 +1,1 @@
+export { OrderDetailsModal } from "./OrderDetailsModal";
